\name{HW5dtreisman-package}
\alias{HW5dtreisman-package}
\alias{HW5dtreisman}
\docType{package}
\title{
\packageTitle{HW5dtreisman}
}
\description{
\packageDescription{HW5dtreisman}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW5dtreisman}
\packageIndices{HW5dtreisman}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HW5dtreisman}

Maintainer: \packageMaintainer{HW5dtreisman}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
